% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_functions.R
\name{aggregate_annual_cat_subcat}
\alias{aggregate_annual_cat_subcat}
\title{Summarize data by year by category within category type OR subcategory within category}
\usage{
aggregate_annual_cat_subcat(dataset, high_cat, low_cat, dataset_all)
}
\arguments{
\item{dataset}{A dataframe of raw lmr data with filters applied}

\item{high_cat}{Character string representing highest level category type or category or subcategory}

\item{low_cat}{Character string representing low level category/subcategory}

\item{dataset_all}{A dataframe of raw unfiltered lmr data for calculating percent share}
}
\value{
A dataframe of aggregated data by year by category_type/category or category/subcategory (up to two levels)
}
\description{
Summarize data by year by category within category type OR subcategory within category
}
