% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_functions.R
\name{aggregate_annual_cat_type}
\alias{aggregate_annual_cat_type}
\title{Summarize data by category type by year}
\usage{
aggregate_annual_cat_type(dataset, dataset_all = beer_data)
}
\arguments{
\item{dataset}{A dataframe of raw lmr data with filters applied}

\item{dataset_all}{A dataframe of raw unfiltered lmr data for calculating percent share}
}
\value{
A dataframe of aggregated data by category type, year
}
\description{
Summarize data by category type by year
}
