% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_functions.R
\name{aggregate_qtr_cat_type}
\alias{aggregate_qtr_cat_type}
\title{Summarize data by category type by quarter}
\usage{
aggregate_qtr_cat_type(dataset, n_qtr = 4)
}
\arguments{
\item{dataset}{A dataframe of raw lmr data with filters applied}

\item{n_qtr}{Number of quarters in each yr; default is 4}
}
\value{
A dataframe of aggregated data by category type, quarter
}
\description{
Summarize data by category type by quarter
}
