% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database_functions.R
\name{conversions}
\alias{conversions}
\title{convert specified fields to factors and create cyr_qtr field}
\usage{
conversions(data)
}
\arguments{
\item{data}{A data frame of lmr data}
}
\value{
A data frame with selected fields converted
}
\description{
convert specified fields to factors and create cyr_qtr field
}
