% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{db_disconnect}
\alias{db_disconnect}
\title{Disconnect from the LMR database
recommended by ChatGPT to put here for easier testing}
\usage{
db_disconnect(con)
}
\description{
Disconnect from the LMR database
recommended by ChatGPT to put here for easier testing
}
\keyword{internal}
