% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{db_get_query}
\alias{db_get_query}
\title{Recommended by ChatGPT for easier test structure}
\usage{
db_get_query(con, query, params)
}
\description{
Recommended by ChatGPT for easier test structure
}
\keyword{internal}
