% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database_functions.R
\name{fetch_db_basic}
\alias{fetch_db_basic}
\title{fetch any lmr (fig4)table data - uses new dbx_get_con function for ease of use
defaults to main data table
use for: quick fetch any raw table
for lmr data with filters: use fetch_LMR_filtered}
\usage{
fetch_db_basic(db_tbl = "public.lmr_data")
}
\arguments{
\item{db_tbl}{Character string of the table name to fetch (default: "public}
}
\value{
A data frame of the requested table data
}
\description{
fetch any lmr (fig4)table data - uses new dbx_get_con function for ease of use
defaults to main data table
use for: quick fetch any raw table
for lmr data with filters: use fetch_LMR_filtered
}
