% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database_functions.R
\name{fetch_lmr_complete_filter}
\alias{fetch_lmr_complete_filter}
\title{fetch complete lmr data with optional filters (using short names)
includes quarter info and short names; optionally, short names can replace originals}
\usage{
fetch_lmr_complete_filter(
  replace = FALSE,
  cat_type_short = NULL,
  category_short = NULL,
  subcategory_short = NULL,
  min_end_qtr_dt = NULL,
  max_end_qtr_dt = NULL
)
}
\arguments{
\item{replace}{Logical - replace original category names with short names}

\item{cat_type_short}{Character string of category type short name (default: NULL)}

\item{category_short}{Character string of category short name (default: NULL)}

\item{subcategory_short}{Character string of subcategory short name(default: NULL)}

\item{min_end_qtr_dt}{date YYYY-MM-DD (default: NULL)}

\item{max_end_qtr_dt}{date YYYY-MM-DD (default: NULL)}
}
\value{
A data frame of the requested table data
}
\description{
fetch complete lmr data with optional filters (using short names)
includes quarter info and short names; optionally, short names can replace originals
}
