% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database_functions.R
\name{fetch_lmr_filter}
\alias{fetch_lmr_filter}
\title{fetch lmr data with optional filters}
\usage{
fetch_lmr_filter(
  cat_type = NULL,
  category = NULL,
  subcategory = NULL,
  min_qtr = NULL,
  max_qtr = NULL
)
}
\arguments{
\item{cat_type}{Character string of category type (default: NULL)}

\item{category}{Character string of category (default: NULL)}

\item{subcategory}{Character string of subcategory (default: NULL)}

\item{min_qtr}{Character string of min qtr as FY20XXQX (default: NULL)}

\item{max_qtr}{Character string of max qtr as FY20XXQX (default: NULL)}
}
\value{
A data frame of the requested table data
}
\description{
fetch lmr data with optional filters
}
