% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_con}
\alias{get_con}
\title{Establish connection to the LMR database}
\usage{
get_con()
}
\value{
A DBI connection object to the LMR database (technically Fig4)
}
\description{
Establish connection to the LMR database
}
\keyword{internal}
