% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database_functions.R
\name{get_max_date}
\alias{get_max_date}
\title{get max date in the LMR database}
\usage{
get_max_date()
}
\value{
A date object of the max end_qtr_dt in the LMR database
}
\description{
get max date in the LMR database
}
