% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database_functions.R
\name{list_tables}
\alias{list_tables}
\title{List the tables in the LMR database}
\usage{
list_tables()
}
\value{
A character vector of table names
}
\description{
List the tables in the LMR database
}
