% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{na_if_missing}
\alias{na_if_missing}
\title{Internal helper for database_functions.R SQL queries
from ChatGPT}
\usage{
na_if_missing(x)
}
\description{
Internal helper for database_functions.R SQL queries
from ChatGPT
}
\keyword{internal}
