% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database_functions.R
\name{fetch_lmr_complete_filter}
\alias{fetch_lmr_complete_filter}
\title{fetch complete lmr data with optional filters
includes quarter info and short names; optionally, short names can replace originals}
\usage{
fetch_lmr_complete_filter(
  replace = FALSE,
  cat_type = NULL,
  category = NULL,
  subcategory = NULL,
  min_end_qtr_dt = NULL,
  max_end_qtr_dt = NULL
)
}
\arguments{
\item{replace}{Logical - replace original category names with short names}

\item{cat_type}{Character string of category type (default: NULL)}

\item{category}{Character string of category (default: NULL)}

\item{subcategory}{Character string of subcategory (default: NULL)}

\item{min_end_qtr_dt}{date YYYY-MM-DD (default: NULL)}

\item{max_end_qtr_dt}{date YYYY-MM-DD (default: NULL)}
}
\value{
A data frame of the requested table data
}
\description{
fetch complete lmr data with optional filters
includes quarter info and short names; optionally, short names can replace originals
}
