% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database_functions.R
\name{get_con}
\alias{get_con}
\title{Establish a connection to the LMR database}
\usage{
get_con()
}
\value{
A DBI connection object to the LMR database (technically Fig4)
}
\description{
Establish a connection to the LMR database
}
